Updates to HLS implementation
The HLS implementation of the merge IP in the initial commit does not work. The
problem being that I did not make sure the performance of the core allowed for a
single cycle latency. Or in other words that a new ADC sample was consumed
every clock cycle.

At one point I had fixed it and tested it on a counter with several designs
showing that it now worked but I never did close the loop on cleaning up those
changes. I think this is the commit that does that but we would need to verify
and double check that is true.